fix FatVC
