+++ Marindeck/View/Core/ViewController+Action.swift
@@ -1,0 +1,16 @@
+//
+//  ViewController+Action.swift
+//  Marindeck
+//
+//  Created by a on 2022/02/05.
+//
+
+import Foundation
+
+protocol ViewControllerAction {
+    
+}
+
+extension ViewController: ViewControllerAction {
+    
+}
