+++ Marindeck/View/Core/ViewCotroller+ViewSetup.swift
@@ -8,7 +8,14 @@
 import UIKit
 import WebKit
 
-extension ViewController {
+protocol ViewSetup {
+    func setupView()
+    func setupTweetBtn()
+    func setupWebViewToolBar()
+    func notchLogoSetup()
+}
+
+extension ViewController: ViewSetup {
 
     func setupView() {
         view.backgroundColor = #colorLiteral(red: 0.08181380481, green: 0.1257319152, blue: 0.1685300171, alpha: 1)
